app/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
Pillow
opencv-python